refactor(runtime): extract provider status helpers
