+++ src/renderer/store/slices/cliInstallerSlice.ts
@@ -18,6 +18,12 @@ import { CLI_PROVIDER_STATUS_DEFERRED_MESSAGE } from '@shared/types/cliInstaller
 import { createLogger } from '@shared/utils/logger';
 import { createDefaultCliExtensionCapabilities } from '@shared/utils/providerExtensionCapabilities';
 
+import {
+  reconcileCliProviderSnapshot,
+  revokeProviderLaunchAuthority,
+  settleCliProviderStatusLoading,
+} from './cliInstallerStatusReconciliation';
+
 import type { AppState } from '../types';
 import type { CodexRuntimeStatus } from '@features/codex-runtime-installer/contracts';
 import type {
@@ -218,116 +224,6 @@ function hasCodexRuntimeReady(provider: CliProviderStatus | undefined): boolean 
     provider?.providerId === 'codex' &&
     provider.availableBackends?.some((backend) => backend.id === 'codex-native') === true
   );
-}
-
-function mergeProviderCatalogCache(
-  incomingProvider: CliProviderStatus,
-  currentProvider: CliProviderStatus
-): CliProviderStatus {
-  const retainedCatalog = incomingProvider.modelCatalog ?? currentProvider.modelCatalog ?? null;
-  const catalogRetained =
-    incomingProvider.modelCatalog == null && currentProvider.modelCatalog != null;
-  const modelEvidenceRetained =
-    incomingProvider.models.length === 0 &&
-    (incomingProvider.modelAvailability?.length ?? 0) === 0 &&
-    !(
-      incomingProvider.modelCatalog?.status === 'ready' &&
-      incomingProvider.modelCatalog.models.length > 0
-    ) &&
-    (currentProvider.models.length > 0 || (currentProvider.modelAvailability?.length ?? 0) > 0);
-  const launchUnproved =
-    incomingProvider.statusCheckOutcome !== 'authoritative' ||
-    incomingProvider.statusCheckErrorCode != null ||
-    (incomingProvider.modelCatalog != null && incomingProvider.modelCatalog.status !== 'ready') ||
-    catalogRetained ||
-    modelEvidenceRetained;
-  const modelCatalog =
-    retainedCatalog && launchUnproved
-      ? { ...retainedCatalog, status: 'stale' as const }
-      : retainedCatalog;
-  return {
-    ...incomingProvider,
-    supported: incomingProvider.supported,
-    authenticated: launchUnproved ? false : incomingProvider.authenticated,
-    authMethod: launchUnproved ? null : incomingProvider.authMethod,
-    canLoginFromUi: launchUnproved
-      ? currentProvider.canLoginFromUi
-      : incomingProvider.canLoginFromUi,
-    capabilities: launchUnproved
-      ? { ...incomingProvider.capabilities, teamLaunch: false }
-      : incomingProvider.capabilities,
-    selectedBackendId: launchUnproved
-      ? currentProvider.selectedBackendId
-      : incomingProvider.selectedBackendId,
-    resolvedBackendId: launchUnproved
-      ? currentProvider.resolvedBackendId
-      : incomingProvider.resolvedBackendId,
-    models: incomingProvider.models.length > 0 ? incomingProvider.models : currentProvider.models,
-    modelAvailability:
-      (incomingProvider.modelAvailability?.length ?? 0) > 0
-        ? incomingProvider.modelAvailability
-        : currentProvider.modelAvailability,
-    availableBackends:
-      (incomingProvider.availableBackends?.length ?? 0) > 0
-        ? incomingProvider.availableBackends
-        : currentProvider.availableBackends,
-    externalRuntimeDiagnostics:
-      (incomingProvider.externalRuntimeDiagnostics?.length ?? 0) > 0
-        ? incomingProvider.externalRuntimeDiagnostics
-        : currentProvider.externalRuntimeDiagnostics,
-    backend: incomingProvider.backend ?? currentProvider.backend,
-    connection: incomingProvider.connection ?? currentProvider.connection,
-    modelCatalog,
-    modelCatalogRefreshState:
-      modelCatalog && launchUnproved
-        ? incomingProvider.modelCatalogRefreshState === 'loading'
-          ? 'loading'
-          : 'error'
-        : (incomingProvider.modelCatalogRefreshState ?? currentProvider.modelCatalogRefreshState),
-    runtimeCapabilities:
-      incomingProvider.runtimeCapabilities ?? currentProvider.runtimeCapabilities ?? null,
-    subscriptionRateLimits:
-      incomingProvider.subscriptionRateLimits ?? currentProvider.subscriptionRateLimits ?? null,
-  };
-}
-
-function revokeProviderLaunchAuthority(provider: CliProviderStatus): CliProviderStatus {
-  return {
-    ...provider,
-    authenticated: false,
-    authMethod: null,
-    capabilities: { ...provider.capabilities, teamLaunch: false },
-    modelCatalog: provider.modelCatalog ? { ...provider.modelCatalog, status: 'stale' } : null,
-    modelCatalogRefreshState: provider.modelCatalog
-      ? provider.modelCatalogRefreshState === 'loading'
-        ? 'loading'
-        : 'error'
-      : provider.modelCatalogRefreshState,
-  };
-}
-
-/** Retains same-provider display evidence without retaining uncertain launch authority. */
-function reconcileCliProviderSnapshot(
-  currentProvider: CliProviderStatus | undefined,
-  incomingProvider: CliProviderStatus
-): CliProviderStatus {
-  if (currentProvider && currentProvider.providerId !== incomingProvider.providerId) {
-    return revokeProviderLaunchAuthority(currentProvider);
-  }
-  const mergedProvider = currentProvider
-    ? mergeProviderCatalogCache(incomingProvider, currentProvider)
-    : incomingProvider;
-  if (
-    incomingProvider.statusCheckOutcome === 'authoritative' &&
-    incomingProvider.statusCheckErrorCode == null &&
-    (incomingProvider.modelCatalog == null || incomingProvider.modelCatalog.status === 'ready') &&
-    (!currentProvider ||
-      (incomingProvider.modelCatalog != null && incomingProvider.models.length > 0) ||
-      (currentProvider.modelCatalog == null && currentProvider.models.length === 0))
-  ) {
-    return mergedProvider;
-  }
-  return revokeProviderLaunchAuthority(mergedProvider);
 }
 
 export function getIncompleteMultimodelProviderIds(
@@ -1408,12 +1304,11 @@ export const createCliInstallerSlice: StateCreator<AppState, [], [], CliInstalle
         }
         set((state) => {
           const currentCliStatus = state.cliStatus;
-          const nextLoading = silent && !projectPath
-            ? state.cliProviderStatusLoading
-            : {
-                ...state.cliProviderStatusLoading,
-                [providerId]: false,
-              };
+          const nextLoading = settleCliProviderStatusLoading(
+            state.cliProviderStatusLoading,
+            providerId,
+            { silent, projectPath }
+          );
 
           if (
             requestEpoch !== cliStatusEpoch ||
@@ -1522,12 +1417,11 @@ export const createCliInstallerSlice: StateCreator<AppState, [], [], CliInstalle
         logger.error(`Failed to fetch ${providerId} CLI status:`, error);
         set((state) => {
           const currentCliStatus = state.cliStatus;
-          const nextLoading = silent && !projectPath
-            ? state.cliProviderStatusLoading
-            : {
-                ...state.cliProviderStatusLoading,
-                [providerId]: false,
-              };
+          const nextLoading = settleCliProviderStatusLoading(
+            state.cliProviderStatusLoading,
+            providerId,
+            { silent, projectPath }
+          );
 
           if (
             requestEpoch !== cliStatusEpoch ||
